[회원관리 > 셀러계정관리](선택셀러 상세보기 창 연동  작업)
