+++ frontend/public/sellerList.json
@@ -15,6 +15,7 @@
       "manager_phone_number": "010-1111-2222 ok",
       "manager_email": "yeni@gg.com ok",
       "number_of_product": 3,
+      "seller_key_id": 900227,
       "site_url": "http://www.chunchun.com",
       "created_at": "2020-06-05 20: 34: 12",
       "actions_by_status": [
