+++ frontend/src/config/urlConfig.js
@@ -1,7 +1,6 @@
-export const URL = "http://10.250.2.85:8080/" //vue local url 주소입니다.
+export const URL = "http://192.168.0.168:8080/" //vue local url 주소입니다.
 // export const SJ_URL = "http://192.168.7.40:5000" // 성준님 1층 url
-export const SJ_URL = "http://10.250.2.195:5000/" // 성준님 2층 url
-export const YE_RUL = "" // 예은님 url
+export const SJ_URL = "http://10.58.3.223:5000/" // 성준님 2층 url
+export const YE_URL = "http://10.58.3.11:5000/" // 예은님 url
 export const JA_URL = "" // 진아님 url
 export const JH_RUL = "" // 진현님 url
-
