+++ frontend/src/views/SellerRegist/SellerRegist02.vue
@@ -1,0 +1,21 @@
+<template>
+  <div>
+    <SrTitle />
+    <BasicInfo />
+  </div>
+</template>
+
+<script>
+import SrTitle from "./Components/SrTitle";
+import BasicInfo from "./Components/BasicInfo";
+
+export default {
+  components: {
+    SrTitle,
+    BasicInfo
+  }
+};
+</script>
+
+<style>
+</style>
